GCBits methods for mass bit modification
Added convenience methods for handling a lot of consecutive bits,
needed for the is_pointer bitmap. These methods are really simple,
and should be replaced with ones optimized for speed. Given that they
run on every alloc and free, SIMD would probably not be overkill.

copyRange copies bits from a provided byte[] until it runs out, and
the clears the rest of the target area

setRange sets len bits starting at target to content